+++ src/discord_client.c
@@ -16,7 +16,7 @@ int discord_login(SSL *ssl, const char *email, const char *password, char **uid,
   if (!body) {
     return 1;
   }
-  if (snprintf(body, body_size + 1, body_fmt, email, password) != body_size) {
+  if (snprintf(body, (size_t)body_size + 1, body_fmt, email, password) != body_size) {
     free(body);
     return 1;
   }
